Add multi-GPU optimization and fix directory structure
- Update Ray workers to use 2 GPUs per worker for improved performance
- Fix device_map issues by using torch_dtype=torch.float16 directly
- Add /data directory symlinks in Ray workers for video output
- Configure imagePullPolicy: Always for all deployments
- Optimize warm model loading with proper GPU utilization
- Enhance Ray worker startup scripts with directory creation

+++ ray_worker_warmup.py
@@ -40,8 +40,11 @@ def preload_mochi_model():
     start_time = time.time()
     
     try:
-        # Load the model
-        _MOCHI_PIPELINE_CACHE = MochiPipeline.from_pretrained("genmo/mochi-1-preview").to("cuda")
+        # Load the model with float16 precision to reduce memory usage
+        _MOCHI_PIPELINE_CACHE = MochiPipeline.from_pretrained(
+            "genmo/mochi-1-preview", 
+            torch_dtype=torch.float16
+        ).to("cuda")
         _MOCHI_PIPELINE_CACHE.enable_model_cpu_offload()
         _MOCHI_PIPELINE_CACHE.enable_vae_tiling()
